maintanance: aligned 34idc and esrf_01 with simple beamline, removed multi_rec, added main in cohere-scripts, moved logic from phasing to simple_phasing.

+++ cohere-scripts/mpi_cmd.py
@@ -14,11 +14,9 @@ def show_log(file):
             print(lines[-1])
 
 
-def run_with_mpi(ga_method, lib, conf_file, datafile, dir, devices, hostfile=None):
+def run_with_mpi(lib, conf_file, datafile, dir, devices, hostfile=None):
     """
 
-    :param ga_method: str
-        info about what kind of case is requested
     :param lib: str
         defines package used in reconstruction
     :param conf_file: str
@@ -35,22 +33,22 @@ def run_with_mpi(ga_method, lib, conf_file, datafile, dir, devices, hostfile=Non
     """
     p = None
     start_time = time.time()
-    if ga_method is None:
-        script = '/reconstruction_multi.py'
-    else:
-        script = '/reconstruction_GA.py'
-        # start process that will monitor log file and print progress
-        log_file = 'ga.log'
-        open(log_file, 'w')
-        p = Process(target=show_log, args=(log_file,))
-        p.start()
+
+    script = '/reconstruction_GA.py'
+    # start process that will monitor log file and print progress
+    log_file = f'ga_{os.getpid()}.log'
+    open(log_file, 'w')
+    p = Process(target=show_log, args=(log_file,))
+    p.start()
 
 
     script = os.path.realpath(os.path.dirname(__file__)).replace(os.sep, '/') + script
     if hostfile is None:
-        command = ['mpiexec', '-n', str(len(devices)), 'python', script, lib, conf_file, datafile, dir, str(devices)]
+        command = ['mpiexec', '-n', str(len(devices)), 'python', script,
+                   lib, conf_file, datafile, dir, str(devices), log_file]
     else:
-        command = ['mpiexec', '-n', str(len(devices)), '--hostfile', hostfile, 'python', script, lib, conf_file, datafile, dir, str(devices)]
+        command = ['mpiexec', '-n', str(len(devices)), '--hostfile', hostfile, 'python', script,
+                   lib, conf_file, datafile, dir, str(devices), log_file]
 
     subprocess.run(command, check=True, capture_output=True)
     run_time = time.time() - start_time
@@ -61,12 +59,10 @@ def run_with_mpi(ga_method, lib, conf_file, datafile, dir, devices, hostfile=Non
         while p.is_alive():
             time.sleep(.1)
         p.close()
-        os.remove(log_file)
+        if os.path.isfile(log_file):
+            os.remove(log_file)
 
-    if ga_method is None:
-        print(f'multiple reconstructions took {run_time} seconds')
-    else:   # fast GA
-        print(f'GA reconstruction took {run_time} seconds')
+    print(f'GA reconstruction took {run_time} seconds')
